yml did not connect to the database, switched to yaml.
################# NOTE FOR FUTURE VERY HELPFUL

Jackson made the models go to recursion and caused infinite loop with an exception, as I read it has to do with serializing. There are 2 annotations, @JsonBackReference for the class which is referencing and @JsonManagedReference for the class which is the "main". Had to add the Jackson and it's annotation dependencies, as I am working on multi module system and don't have Jackson installed.